requirements.txt
requests==2.28.1
flask==2.1.1
pytest==7.1.2
pyyaml